Update Thunderbird announcement description field
Modify the description field to be more of a general note
that inherited Firefox advisories may not affect Thunderbird.
The previous wording may have given the impression that none
of the advisories affected Thunderbird.

With advisories from Firefox and Thunderbird mixed together,
this will help convey to users that the advisories may affect
Thunderbird.