Fix possibly undefined "references" key
This didn't used to be a problem, but recently, some advisories seem to
fail to be published, because the containers>cna>references field
doesn't exist in the JSON coming from CVE services. I suspect this may
be happening because there just are no references for that CVE yet,
which CVE services report by just not setting that field at all. We
previously probably just never had the situation that an existing CVE
has no references because we almost always add at least a Bugzilla link.

As a fix, just check if the "references" key exists beforehand.